feat(rust): added skip serde deserialize attr for tags in `servicelist`, nodeopts to list services
