+++ implementations/rust/ockam/ockam_command/src/service/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod config;
 pub(crate) mod list;
 pub(crate) mod start;
+pub(crate) mod util;
 
 pub(crate) use start::StartCommand;
