+++ implementations/rust/ockam/ockam_command/src/util/output.rs
@@ -3,7 +3,6 @@ use cli_table::{Cell, Style, Table};
 use core::fmt::Write;
 use ockam::identity::credential::Credential;
 use ockam_api::cloud::project::{Enroller, Project};
-use ockam_api::nodes::models::services::ServiceList;
 
 use crate::project::ProjectInfo;
 use crate::util::comma_separated;
@@ -235,25 +234,5 @@ impl Output for Vec<Enroller<'_>> {
 impl Output for Credential<'_> {
     fn output(&self) -> anyhow::Result<String> {
         Ok(self.to_string())
-    }
-}
-
-impl Output for ServiceList<'_> {
-    fn output(&self) -> anyhow::Result<String> {
-        if self.list.is_empty() {
-            return Ok("No services found".to_string());
-        }
-
-        let mut w = String::new();
-        write!(w, "Services:")?;
-
-        let services_list = self.list.clone();
-        for service in services_list {
-            write!(w, "\n  Service: ")?;
-            write!(w, "\n    Type: {}", service.service_type)?;
-            write!(w, "\n    Address: /service/{}", service.addr)?;
-        }
-
-        Ok(w)
     }
 }
